test(types): expect suint length for dynamic shielded arrays
Update all tests involving dynamic shielded arrays to expect .length
to return suint256 instead of uint256, and to use cload/cstore instead
of sload/sstore for the length slot in inline assembly.

Add new semantic and syntax tests for comprehensive coverage of the
suint length behavior including type checking, cload/cstore assembly
verification, push/pop operations, and mapping interactions.

+++ test/libsolidity/semanticTests/types/mapping/shielded_copy_from_mapping_to_mapping.sol
@@ -34,9 +34,9 @@ contract C {
     }
 
     function toUnshielded(S memory s) internal pure returns (SUnshielded memory) {
-        uint8[][] memory y = new uint8[][](s.y.length);
-        for (uint256 i = 0; i < s.y.length; i++) {
-            uint256 innerLen = s.y[i].length;
+        uint8[][] memory y = new uint8[][](uint256(s.y.length));
+        for (uint256 i = 0; i < uint256(s.y.length); i++) {
+            uint256 innerLen = uint256(s.y[i].length);
             y[i] = new uint8[](innerLen);
             for (uint256 j = 0; j < innerLen; j++) {
                 y[i][j] = uint8(s.y[i][j]);
